Move payment reference calculator
